refactor: http client and service
