+++ src/entities/IContact.ts
