+++ src/lib/utils.ts
@@ -28,7 +28,7 @@ export function generateId() {
   return Math.floor(Math.random() * (max - min + 1)) + min
 }
 
-export function delay(ms: number = 500) {
+export function sleep(ms: number = 500) {
   return new Promise((resolve) => {
     setTimeout(resolve, ms)
   })
