+++ src/services/contacts-service.ts
@@ -1,31 +1,11 @@
-import { AxiosRequestConfig, AxiosResponse } from 'axios'
+import { AxiosRequestConfig } from 'axios'
 
-import { IContact } from '@/interfaces/IContact'
+import { IContact } from '@/entities/IContact'
 
-import { HttpClient } from './utils/http-client'
+import { httpClient } from './utils/http-client'
 
-interface IContactsService {
-  listContacts(config: AxiosRequestConfig): Promise<AxiosResponse<IContact[]>>
-}
-
-class ContactsService implements IContactsService {
-  private httpClient: HttpClient
-
-  constructor() {
-    this.httpClient = new HttpClient({
-      baseURL: 'https://my.api.mockaroo.com/lestetelecom/test.json',
-      params: {
-        key: 'f55c4060',
-      },
-    })
-  }
-
-  public async listContacts(
-    config: AxiosRequestConfig,
-  ): Promise<AxiosResponse<IContact[]>> {
-    const response = await this.httpClient.get<IContact[]>(config)
-    return response
+export class ContactsService {
+  static async listContacts(config?: AxiosRequestConfig) {
+    return await httpClient.get<IContact[]>('', config)
   }
 }
-
-export default new ContactsService()
